requirements.txt
beautifulsoup4
requests
PyMuPDF
langdetect
selenium
webdriver-manager
fastapi
uvicorn[standard]
python-multipart
reportlab
fpdf2
streamlit  # optionnel si tu veux une interface